% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{batch_converter}
\alias{batch_converter}
\title{batch community converter}
\usage{
batch_converter(graph_batch, graph_ref)
}
\arguments{
\item{graph_batch}{: a list of graph to convert}

\item{graph_ref}{: the reference used for conversion}
}
\description{
convert community (give the same indice) of a batch
of graphs to the community of a reference graph
}
