% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso_analysis.R
\name{community_contrast_dashboard}
\alias{community_contrast_dashboard}
\title{community_contrast_dashboard}
\usage{
community_contrast_dashboard(community_table, contrast_name, contrast_vector,
  nrow = 4, levels_class = c("low", "high"), color_vect = NULL,
  offset = 10^-8)
}
\arguments{
\item{community_table}{: a table of community composition generated by the Compute_community_abondance function}

\item{contrast_name}{: how the contrast column should be named}

\item{contrast_vector}{: the contrast vector}
}
\description{
create a graphic with a facet by community and a boxplot by modality of the contrast vector
}
