% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{community_converter}
\alias{community_converter}
\title{community_converter}
\usage{
community_converter(nodes_graph1, nodes_graph2, join_type = "inner")
}
\arguments{
\item{nodes_graph1}{is the table of nodes with the walktrap column properly filled}

\item{nodes_graph2}{is the table of nodes of the second graph}
}
\description{
Community converter function take to two nodes table as argument
and give a translation of each community to the other graph based on jacquart distance
If I may, it is automated community translation
}
