% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso.R
\name{Compute_community_abondance}
\alias{Compute_community_abondance}
\title{compute community abundances}
\usage{
Compute_community_abondance(Nodes, abundance, species_taxo, contrast = NULL,
  community_kind = "walktrap_community")
}
\arguments{
\item{Nodes:}{The network node table with the a community attribution column}

\item{abundance:}{The Metagenomic species abundance table}

\item{species_taxo:}{The Metagenomic species taxonomy table}

\item{contrast:}{a boolean vector to form two group of samples. for each community the rank test difference of abundance p-value is calculated between the two groups.}

\item{community}{kind: the algorithm of used to compute the community : "spinglass_community", "walktrap_community"}
}
\value{
a table of community abundance and composition
}
\description{
Compute the sum of abundance of species in each community given the abundance matrix by species and the graph community
Also compute the p-value of difference of abundance given a contrast vector with to class (2 classes)
}
