% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso.R
\name{Compute_graph}
\alias{Compute_graph}
\title{Compute Glasso}
\usage{
Compute_graph(MGS_abundance, contrast_vector, fout,
  abundance_treshold = 10^-7, occurence_treshold = 0.05, nlambda = 20,
  lambda.min.ratio = 0.1, lambda = NULL, rep.num = 20)
}
\arguments{
\item{contrast_vector}{: a boolean vector to select a subset of the cohort (the model will be infered on samples with a TRUE value)}

\item{MGS_abundance:}{The Metagenomic species abundance table}

\item{fout:}{where to save the model object}

\item{community:}{Should the community structure be calculated?}

\item{nlambda:}{Number of regularisation parameter that will be tested (see huge::huge() documentation)}

\item{lambda.min.ratio:}{the smallest value of lambda as a fraction of its maximum (see huge::huge() documentation)}

\item{occurence_treshold:}{minimum fraction of samples where a species must be present to be taken into account in the analysis}

\item{abundance_threshold:}{minimum mean abundances for a species to be included in the analysis}

\item{rep.num:}{Number of subsampling to compute the edge stability with "StarS" (see huge::huge.select documentation)}

\item{lambda:}{A sequence of regularisation parameter. If not null, it will override the automatic computation of the lambda sequence (with nlambda and lambda.min.ratio)}
}
\description{
Compute the Glasso model from the MGS abundances on individuals with a true value in the
contrast vector.
}
