% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{concordance_table}
\alias{concordance_table}
\title{concordance_table}
\usage{
concordance_table(nlist, Graph_tags, join_type = "outer")
}
\arguments{
\item{nlist}{: list of graphs nodes table with the walktrap_community information Available}

\item{Graph_tags}{: list of graph labels}

\item{join_type}{: how to join graph row (outer joins or inner join). Outer join means that all species present in at least one graph will be taken into account. Inner join means that only species present in all graphs will treated.}
}
\description{
Enable one to assess if community found in diverse cohort are the same
Used to generate the alluvial plot
}
