% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso.R
\name{create_graph}
\alias{create_graph}
\title{Create a gephi format graph from the graphical Lasso  model}
\usage{
create_graph(file_input, file_output, MGS_by_taxo_species, species_taxo,
  nspins = 20, mod_rep = 10, community = FALSE, additional_info = NULL,
  spinglass_opt = FALSE, variability_treshold = NULL)
}
\arguments{
\item{spinglass_opt}{: Should the number of spin of the spinglass community be optimized?}

\item{file_input:}{emplacement of the GLASSO model object}

\item{file_output:}{where to save the network representation file}

\item{MGS_by_taxo_species:}{The Metagenomic species abundance table}

\item{species_taxo:}{The Metagenomic species taxonomy table}

\item{nspins:}{number of spin for the spinglass community detection algorithm}

\item{community:}{Should the community structure be calculated?}

\item{additional_info:}{a vector or data.frame containing information to add to the nodes table of the network}

\item{variability_treshold:}{The maximum mean variability for graph edge presence. If null, the optimal covariance matrix will correspond to a variability of 0.05}
}
\description{
The Function also compute the community structure of the graph with various algorithms (betweeness community, walktrap community...)
community specified by the user
}
