% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{Distance_to_community}
\alias{Distance_to_community}
\title{##################################################
Function Distance to community
Compute mean distance of species to all community}
\usage{
Distance_to_community(my_dist, Nodes_with_com)
}
\arguments{
\item{my_dist}{: distance matrix of species to all species}

\item{Nodes_with_com:}{graph node table containing walktrap community
###################################################}
}
\description{
##################################################
Function Distance to community
Compute mean distance of species to all community
}
