% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso_analysis.R
\name{draw_community_total_abundance}
\alias{draw_community_total_abundance}
\title{draw community total abundance treshold}
\usage{
draw_community_total_abundance(abund, color_map = NULL)
}
\arguments{
\item{abund}{: table of community abundance accross samples}
}
\description{
A function to draw a barplot of the mean abundance of community accross samples
}
