% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso_analysis.R
\name{draw_community_total_species_count}
\alias{draw_community_total_species_count}
\title{draw_community_total_species_count}
\usage{
draw_community_total_species_count(community_table)
}
\arguments{
\item{community_table}{: Community table with the species composition line}
}
\description{
draw a barplot of the number of species by community
}
