% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso_analysis.R
\name{extract_community_abundance_table}
\alias{extract_community_abundance_table}
\title{extract_community_abundance_table}
\usage{
extract_community_abundance_table(Community_table)
}
\arguments{
\item{Community_table}{: a table of community composition generated by the Compute_community_abondance function}
}
\description{
function to handily extract only the abondance of community by
samples tables (and remove the extra information from the table like community composition)
}
