% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso.R
\name{Gleesso_pipeline}
\alias{Gleesso_pipeline}
\title{Pipeline launcher}
\usage{
Gleesso_pipeline(data_folder, MGS_file, taxo_file, model_folder, graph_folder,
  contrast_vector, tag_model, tag_graph, community = TRUE, nlambda = 20,
  lambda.min.ratio = 0.1, occurence_treshold = 0.05,
  abundance_treshold = 10^-7, variability_treshold = NULL,
  analysis_step = NULL, species_mode = TRUE)
}
\arguments{
\item{data_folder}{: where the community abundance table will be written}

\item{contrast_vector}{: a boolean vector to select a subset of the cohort (the model will be infered on samples with a TRUE value)}

\item{tag_model}{: a tag that will be inserted in output file to recognize the model parameter}

\item{tag_graph}{: a tag that will be inserted in output file to recognize the graph}

\item{species_mode}{: should the graph inference be done on MGS (FALSE) or with MGS of the same specied merged togethere (TRUE)}

\item{MGS_file:}{The Metagenomic species abundance file in the RDS format}

\item{model_folder:}{where to save the graphical Lasso model object (output of the spiec.easi function)}

\item{graph_folder:}{where the graphical representation of the model will be saved (in the gephi format)}

\item{community:}{Should the community structure be calculated?}

\item{nlambda:}{Number of regularisation parameter that will be tested (see huge::huge() documentation)}

\item{lambda.min.ratio:}{the smallest value of lambda as a fraction of its maximum (see huge::huge() documentation)}

\item{occurence_treshold:}{minimum fraction of samples where a species must be present to be taken into account in the analysis}

\item{abundance_threshold:}{minimum mean abundances for a species to be included in the analysis}

\item{variability_treshold:}{The maximum mean variability for graph edge presence. If null, the optimal covariance matrix will correspond to a variability of 0.05}

\item{analysis_step:}{At which step the analysis should be started (0: from scratch, 1: model inferences, 2: save gephi network, 3: Community detection). If NULL (default), the step will be infered from the files present in the output folders. Use analysis_step=0 to force computation from scratch.}
}
\description{
Launching the complete glasso analysis (data prep,  graph inference, community detection, community abundances computation)
Warning : folder shouldn't be indicated with an / at the end
}
