% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gleesso.R
\name{opt_spinglass_com}
\alias{opt_spinglass_com}
\title{Computing spinglass communities and their modularity for a range of number of spin
 We then retrieve the optimal number of spin according to modularity}
\usage{
opt_spinglass_com(con.grph, mod_rep, nspins)
}
\description{
Computing spinglass communities and their modularity for a range of number of spin
 We then retrieve the optimal number of spin according to modularity
}
