% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{parrallel_coord_community}
\alias{parrallel_coord_community}
\title{#######################################################
Draw a parrallel coord graph of community belonging for
different graph object. Enable one to compare and understand the stability or discrepancy
between graph community}
\usage{
parrallel_coord_community(graph_node_list, Graph_tags, measure = "sum_ab",
  color_graph = 1, join_type = "inner")
}
\arguments{
\item{color_graph}{: index of the graph used to color the parallel coordiante plot
###################################}

\item{graph_node_list:}{a sequence of nodes tables with community annotated}

\item{Graph_tags:}{a sequence of str which are the name of nodes tables of graph_node_list}

\item{measure:}{the weight attributed to each CAG either "sum" of abundance or "count" of objects}
}
\description{
#######################################################
Draw a parrallel coord graph of community belonging for
different graph object. Enable one to compare and understand the stability or discrepancy
between graph community
}
