% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{Robust_table_community}
\alias{Robust_table_community}
\title{Community attribution stability table}
\usage{
Robust_table_community(graphs_folder, alluvial_diagnostic_file, taxo,
  N_alluvial = 10, join_type = "outer", stability_treshold = 0.6,
  silhouette_treshold = 0.1)
}
\arguments{
\item{graphs_folder}{: folder where all graphs are placed (with the / at the end ^^)}

\item{N_alluvial}{: number of graph to represent on a graph}

\item{alluvial_diagnostic}{: file name for the alluvial graph}
}
\description{
Compute the stability of
species community attribution from bootstraped graphs
}
