% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{Silhouette_to_community}
\alias{Silhouette_to_community}
\title{Silhouette_to_community}
\usage{
Silhouette_to_community(my_dist, Nodes_with_com)
}
\arguments{
\item{my_dist}{: distance matrix of species to all species}

\item{Nodes_with_com}{: Nodes table with a walktrap_community attribution}
}
\description{
Compute the silhouette cluster metric for all species
}
