% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{walktrap_distance}
\alias{walktrap_distance}
\title{function walktrap_distance reproduce the distance used in the walktrap community detection algorithm}
\usage{
walktrap_distance(pos.grph, n_steps)
}
\arguments{
\item{pos.graph:}{an igraph object that contain only positive edges}

\item{n_steps:}{number of steps of the random walk on the graph
##############################################}
}
\description{
function walktrap_distance reproduce the distance used in the walktrap community detection algorithm
}
