% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{Robust_table_community}
\alias{Robust_table_community}
\title{Community attribution stability table}
\usage{
Robust_table_community(graphs_folder, alluvial_diagnostic_file, taxo,
  N_alluvial = 10, join_type = "outer", stability_treshold = 0.6,
  silhouette_treshold = 0.1)
}
\arguments{
\item{graphs_folder}{: folder where all graphs are placed (with the / at the end please ^^)}

\item{N_alluvial}{: number of graph to represent on a graph}

\item{join_type}{: Should we work on the union of species or the intersection}

\item{alluvial_diagnostic}{: file name for the alluvial graph}
}
\value{
a list object containing the following elements:
staby : table of species stability with community assignation for all graphs
stab_n_taxo : table of species stability and taxonomy
Robust_community_stability_[...] : abundance of communities for species with a stability above the specified treshold
Robust_community_stability_[...]_silhouette_[...] : idem but species also have a silhouette above a specified treshold
}
\description{
Compute the stability of
species community attribution from bootstraped graphs and community
table abondance for robust attribution.
}
