% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_community.R
\name{Silhouette_to_community}
\alias{Silhouette_to_community}
\title{cl_community = Nodes_with_com$walktrap_community
 converter = as.data.frame(t(as.data.frame(strsplit(unique(paste(Nodes_with_com$walktrap_community, cl_community, sep="_")), "_"))), stringAsFactors = FALSE)
 row.names(converter) = converter$V2
S_distdt[,1] = converter[SDist[,1], 1]
S_distdt[,2] = converter[SDist[,2], 1]}
\usage{
Silhouette_to_community(my_dist, Nodes_with_com)
}
\description{
cl_community = Nodes_with_com$walktrap_community
 converter = as.data.frame(t(as.data.frame(strsplit(unique(paste(Nodes_with_com$walktrap_community, cl_community, sep="_")), "_"))), stringAsFactors = FALSE)
 row.names(converter) = converter$V2
S_distdt[,1] = converter[SDist[,1], 1]
S_distdt[,2] = converter[SDist[,2], 1]
}
